+++ src/specVersion.ts
@@ -1,1 +1,1 @@
-export const SOUSTACK_SPEC_VERSION = '0.2.1';
+export const SOUSTACK_SPEC_VERSION = '0.3.0';